requirements.txt
dnspython==2.4.2
http3==0.6.7
